Correctly handle negative seek offset in the SE_Descriptor
A negative offset should prevent the initial seek in the sftp layer.

+++ lib-src/rpc2/rpc2-src/sftp3.c
@@ -188,7 +188,7 @@ int sftp_InitIO(struct SFTP_Entry *sEntry)
 	return(-1);
     }
 
-    if (sftpd->SeekOffset >= 0) {
+    if ((int32_t)sftpd->SeekOffset >= 0) {
 	sEntry->fd_offset = (off_t)sftpd->SeekOffset;
 	(void)lseek(sEntry->openfd, sEntry->fd_offset, SEEK_SET);
     }
